+++ java/vector/src/test/java/org/apache/arrow/vector/file/TestArrowFile.java
@@ -17,21 +17,17 @@
  */
 package org.apache.arrow.vector.file;
 
-import static com.google.common.base.Preconditions.checkArgument;
-
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 
 import org.apache.arrow.memory.BufferAllocator;
 import org.apache.arrow.memory.RootAllocator;
-import org.apache.arrow.vector.FieldVector;
-import org.apache.arrow.vector.ValueVector.Accessor;
+import org.apache.arrow.vector.VectorLoader;
+import org.apache.arrow.vector.VectorUnloader;
 import org.apache.arrow.vector.complex.MapVector;
 import org.apache.arrow.vector.complex.impl.ComplexWriterImpl;
 import org.apache.arrow.vector.complex.impl.SingleMapReaderImpl;
@@ -40,15 +36,10 @@ import org.apache.arrow.vector.complex.writer.BaseWriter.ComplexWriter;
 import org.apache.arrow.vector.complex.writer.BaseWriter.MapWriter;
 import org.apache.arrow.vector.complex.writer.BigIntWriter;
 import org.apache.arrow.vector.complex.writer.IntWriter;
-import org.apache.arrow.vector.schema.ArrowFieldNode;
 import org.apache.arrow.vector.schema.ArrowRecordBatch;
-import org.apache.arrow.vector.schema.VectorLayout;
-import org.apache.arrow.vector.types.pojo.Field;
 import org.apache.arrow.vector.types.pojo.Schema;
 import org.junit.Assert;
 import org.junit.Test;
-
-import io.netty.buffer.ArrowBuf;
 
 public class TestArrowFile {
   static final BufferAllocator allocator = new RootAllocator(Integer.MAX_VALUE);
@@ -76,14 +67,13 @@ public class TestArrowFile {
       parent.close();
     }
 
-    System.out.println(file.length());
     {
       try (
           FileInputStream fileInputStream = new FileInputStream(file);
           ArrowReader arrowReader = new ArrowReader(fileInputStream.getChannel(), allocator)
           ) {
         ArrowFooter footer = arrowReader.readFooter();
-        org.apache.arrow.vector.types.pojo.Schema schema = footer.getSchema();
+        Schema schema = footer.getSchema();
         System.out.println("reading schema: " + schema);
 
         // initialize vectors
@@ -92,44 +82,13 @@ public class TestArrowFile {
         MapWriter rootWriter = writer.rootAsMap();
         MapVector root = (MapVector)parent.getChild("root");
 
-        List<Field> fields = schema.getFields();
-        root.initializeChildren(fields);
-        List<FieldVector> fieldVectors = root.getFieldVectors();
-        if (fieldVectors.size() != fields.size()) {
-          throw new IllegalArgumentException(); //TODO
-        }
+        VectorLoader vectorLoader = new VectorLoader(schema, root);
 
-//        validateLayout(fields, parent);
         List<ArrowBlock> recordBatches = footer.getRecordBatches();
         for (ArrowBlock rbBlock : recordBatches) {
           ArrowRecordBatch recordBatch = arrowReader.readRecordBatch(rbBlock);
 
-          Iterator<ArrowBuf> buffers = recordBatch.getBuffers().iterator();
-          Iterator<ArrowFieldNode> nodes = recordBatch.getNodes().iterator();
-          System.out.println(recordBatch.getNodes().size() + " nodes");
-          System.out.println(recordBatch.getBuffers().size() + " buffers");
-          for (int i = 0; i < fields.size(); ++i) {
-            Field field = fields.get(i);
-            FieldVector fieldVector = fieldVectors.get(i);
-            loadBuffers(fieldVector, field, buffers, nodes);
-          }
-
-//          public void load(List<Field> fields, int length, Iterator<ArrowFieldNode> nodes, Iterator<ArrowBuf> buffers) {
-//
-//            for (Field field : fields) {
-//              MinorType minorType = Types.getMinorTypeForArrowType(field.getType());
-//              ValueVector vector = this.add(field.getName(), minorType);
-//
-//
-//              vector.loadBuffers(typeLayout, ownBuffers);
-//              List<Field> children = field.getChildren();
-//              for (Field child : children) {
-//                addChild((NestedVector)vector, child, nodes, buffers);
-//                vector.loadChild()
-//              }
-//            }
-//          }
-//          parent.load(fields, recordBatch.getLength(), nodes, buffers);
+          vectorLoader.load(recordBatch);
 
           MapReader rootReader = new SingleMapReaderImpl(parent).reader("root");
           for (int i = 0; i < count; i++) {
@@ -138,31 +97,10 @@ public class TestArrowFile {
             Assert.assertEquals(i, rootReader.reader("bigInt").readLong().longValue());
           }
 
-          parent.close();
         }
+        parent.close();
       }
 
-    }
-  }
-
-  private void loadBuffers(FieldVector vector, Field field, Iterator<ArrowBuf> buffers, Iterator<ArrowFieldNode> nodes) {
-    ArrowFieldNode fieldNode = nodes.next();
-    List<VectorLayout> typeLayout = field.getTypeLayout().getVectors();
-    List<ArrowBuf> ownBuffers = new ArrayList<>(typeLayout.size());
-    for (int j = 0; j < typeLayout.size(); j++) {
-      ownBuffers.add(buffers.next());
-    }
-    vector.loadFieldBuffers(fieldNode, ownBuffers);
-    List<Field> children = field.getChildren();
-    if (children.size() > 0) {
-      List<FieldVector> childrenFromFields = vector.getChildrenFromFields();
-      int i = 0;
-      checkArgument(children.size() == childrenFromFields.size(), "should have as many children as in the schema: found " + childrenFromFields.size() + " expected " + children.size());
-      for (Field child : children) {
-        FieldVector fieldVector = childrenFromFields.get(i);
-        loadBuffers(fieldVector, child, buffers, nodes);
-        ++i;
-      }
     }
   }
 
@@ -184,40 +122,16 @@ public class TestArrowFile {
 //  }
 
   private void write(MapVector parent, File file) throws FileNotFoundException, IOException {
-    Field rootField = parent.getField();
-    Schema schema = new Schema(rootField.getChildren());
+    VectorUnloader vectorUnloader = new VectorUnloader(parent);
+    Schema schema = vectorUnloader.getSchema();
     System.out.println("writing schema: " + schema);
     try (
         FileOutputStream fileOutputStream = new FileOutputStream(file);
         ArrowWriter arrowWriter = new ArrowWriter(fileOutputStream.getChannel(), schema)
             ) {
-      List<ArrowFieldNode> nodes = new ArrayList<>();
-      List<ArrowBuf> buffers = new ArrayList<>();
-      for (FieldVector vector : parent.getFieldVectors()) {
-        appendNodes(vector, nodes, buffers);
-      }
-      arrowWriter.writeRecordBatch(new ArrowRecordBatch(parent.getAccessor().getValueCount(), nodes, buffers));
+      arrowWriter.writeRecordBatch(vectorUnloader.getRecordBatch());
     }
   }
-
-  private void appendNodes(FieldVector vector, List<ArrowFieldNode> nodes, List<ArrowBuf> buffers) {
-    Accessor accessor = vector.getAccessor();
-    int nullCount = 0;
-    // TODO: should not have to do that
-    // we can do that a lot more efficiently (for example with Long.bitCount(i))
-    for (int i = 0; i < accessor.getValueCount(); i++) {
-      if (accessor.isNull(i)) {
-        nullCount ++;
-      }
-    }
-    nodes.add(new ArrowFieldNode(accessor.getValueCount(), nullCount));
-    // TODO: validate buffer count
-    buffers.addAll(vector.getFieldBuffers());
-    for (FieldVector child : vector.getChildrenFromFields()) {
-      appendNodes(child, nodes, buffers);
-    }
-  }
-
 
 
 }
